requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=2.1.0=pyhd8ed1ab_0
alsa-lib=1.2.10=hd590300_0
aom=3.5.0=h27087fc_0
attr=2.5.1=h166bdaf_1
blas=1.0=mkl
bokeh=3.3.4=pypi_0
bottleneck=1.3.5=py311hbed6279_0
box2d-py=2.3.8=py311hb755f60_8
brotli=1.0.9=h9c3ff4c_4
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.2.2=hbcca054_0
cairo=1.16.0=hb05425b_5
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py311h06a4308_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.2.0=py311hdb19cb5_0
cycler=0.12.1=pyhd8ed1ab_0
dask=2024.2.0=pypi_0
dask-core=2023.11.0=py311h06a4308_0
dav1d=1.2.1=hd590300_0
dbus=1.13.6=h5008d03_3
decorator=4.4.2=pypi_0
distributed=2024.2.0=pypi_0
expat=2.5.0=hcb278e6_1
farama-notifications=0.0.4=pyhd8ed1ab_0
filelock=3.13.1=py311h06a4308_0
fire=0.5.0=pyhd8ed1ab_0
fluidsynth=2.3.4=h16f8cd4_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.1=h4c34cd2_2
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h267a509_2
fsspec=2024.2.0=pypi_0
geos=3.8.0=he6710b0_0
gettext=0.21.1=h27087fc_0
gin-config=0.5.0=pypi_0
glfw=3.3.9=hd590300_0
glib=2.78.3=hfc55251_0
glib-tools=2.78.3=hfc55251_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
graphite2=1.3.13=h58526e2_1001
gymnasium=0.28.1=py311h781c19f_0
gymnasium-box2d=0.28.1=py311h38be061_0
gymnasium-mujoco=0.28.1=py311h38be061_0
harfbuzz=7.3.0=hdb3a94d_0
heapdict=1.0.1=pyhd3eb1b0_0
icu=72.1=hcb278e6_0
idna=3.6=pypi_0
imageio=2.34.0=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
importlib-metadata=7.0.1=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jack=1.9.22=h7c63dc7_2
jax-jumpy=1.0.0=pyhd8ed1ab_0
jinja2=3.1.3=py311h06a4308_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h166bdaf_1
kiwisolver=1.4.4=py311h6a678d5_0
lame=3.100=h166bdaf_1003
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20230125.0=cxx17_hcb278e6_1
libavif=0.11.1=h8182462_2
libcap=2.69=h0f662aa_0
libccd-double=2.1=h59595ed_3
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.17=h5eee18b_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libflac=1.4.3=h59595ed_0
libgcc-ng=13.2.0=h807b86a_5
libgcrypt=1.10.3=hd590300_0
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libglib=2.78.3=h783c2da_0
libgpg-error=1.47=h71f35ed_0
libiconv=1.17=hd590300_2
libllvm14=14.0.6=hdb19cb5_3
libmad=0.15.1b=h0b41bf4_1001
libmujoco=2.3.3=h5c724c8_0
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h5eee18b_0
libsndfile=1.2.2=hc60ed4a_1
libstdcxx-ng=13.2.0=h7e041cc_5
libsystemd0=255=h3516f8a_0
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h0b41bf4_0
libxml2=2.10.4=hfdac1af_0
libzlib=1.2.13=hd590300_5
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.42.0=py311h6a678d5_0
locket=1.0.0=py311h06a4308_0
lodepng=20220109=h924138e_0
lz4=4.3.3=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.5=pypi_0
matplotlib-base=3.8.0=py311ha02d727_0
minigrid=2.3.1=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
moviepy=1.0.3=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpg123=1.32.4=h59595ed_0
mpmath=1.3.0=py311h06a4308_0
msgpack=1.0.7=pypi_0
msgpack-python=1.0.3=py311hdb19cb5_0
mujoco-python=2.3.3=py311h67b707b_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py311h06a4308_0
numba=0.59.0=py311h96b013e_1
numexpr=2.8.7=py311h65dcdc2_0
numpy=1.26.4=pypi_0
numpy-base=1.26.3=py311hf175353_0
numpy_groupies=0.10.2=pyhd8ed1ab_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_0
opusfile=0.12=h3358134_2
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
partd=1.4.1=py311h06a4308_0
pcre=8.45=h295c915_0
pcre2=10.42=hcad00b1_0
pillow=10.2.0=py311h5eee18b_0
pip=23.3.1=py311h06a4308_0
pixman=0.43.2=h59595ed_0
portaudio=19.6.0=h7c63dc7_9
portmidi=2.0.4=h7c63dc7_2
proglog=0.1.10=pypi_0
psutil=5.9.8=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio-client=17.0=hb77b528_0
pyarrow=15.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11-abi=4=hd8ed1ab_3
pygame=2.5.2=pypi_0
pyglfw=2.6.5=py311h38be061_0
pyopengl=3.1.6=pyhd8ed1ab_1
pyparsing=2.4.7=pyhd8ed1ab_1
pyribs=0.7.0=pyhd8ed1ab_1
pyribs-visualize=0.7.0=pyhd8ed1ab_1
python=3.11.7=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-lmdb=1.4.1=py311h6a678d5_0
python-tzdata=2023.4=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytorch=2.2.0=py3.11_cpu_0
pytorch-mutex=1.0=cpu
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py311h5eee18b_0
qhull=2020.2=h4bd325d_2
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
scikit-learn=1.2.2=py311h6a678d5_1
scipy=1.11.4=py311h08b1b3b_0
sdl2=2.28.5=hdbcbe63_1
sdl2_image=2.6.3=h7b3d2e9_0
sdl2_mixer=2.6.3=h8830914_1
sdl2_ttf=2.20.2=h2123e47_1
setuptools=68.2.2=py311h06a4308_0
shapely=2.0.1=py311h24a2a10_0
six=1.16.0=pyh6c4a22f_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sqlite=3.41.2=h5eee18b_0
swig=4.0.2=h295c915_4
sympy=1.12=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tblib=3.0.0=pypi_0
termcolor=2.4.0=pyhd8ed1ab_0
threadpoolctl=3.2.0=pyha21a80b_0
tinyxml2=9.0.0=h9c3ff4c_2
tk=8.6.12=h1ccaba5_0
toolz=0.12.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pyhd8ed1ab_0
typing_extensions=4.9.0=py311h06a4308_1
tzdata=2024.1=pypi_0
urllib3=2.2.0=pypi_0
wheel=0.41.2=py311h06a4308_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.8.7=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxcursor=1.2.0=h0b41bf4_1
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxinerama=1.1.5=h27087fc_0
xorg-libxrandr=1.5.2=h7f98852_1
xorg-libxrender=0.9.11=hd590300_0
xorg-randrproto=1.5.0=h7f98852_1001
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xyzservices=2023.10.1=pypi_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zict=3.0.0=py311h06a4308_0
zipp=3.17.0=py311h06a4308_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hc292b87_0
